one file contract
